+++ migrations/3_orders.up.sql
@@ -1,10 +1,10 @@
 CREATE TABLE orders (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     order_uid TEXT UNIQUE NOT NULL,
     track_number TEXT NOT NULL,
     entry TEXT NOT NULL,
-    delivery_id INT REFERENCES delivery(id),
-    payment_id INT REFERENCES payment(id),
+    delivery_id BIGINT REFERENCES delivery(id),
+    payment_id BIGINT REFERENCES payment(id),
     locale VARCHAR(5) NOT NULL,
     internal_signature TEXT,
     customer_id TEXT,
